organised components into folders and updated references
