+++ src/App.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Nav from './components/common/Nav/Nav';
 import './App.css'
-import Home from './components/Home';
+import Home from './components/Home/Home';
 import GoogleSheets from './components/GoogleSheets';
 import AirTable from './components/AirTable';
 import Footer from './components/common/Footer/Footer'
